rename flowcean/data to flowcean/environments
