+++ docs/user_guide/experiment.md
@@ -49,8 +49,8 @@ import logging
 
 import flowcean.cli
 import polars as pl
-from flowcean.data.dataset import Dataset
-from flowcean.data.train_test_split import TrainTestSplit
+from flowcean.environments.dataset import Dataset
+from flowcean.environments.train_test_split import TrainTestSplit
 from flowcean.learners.linear_regression import LinearRegression
 from flowcean.metrics import MeanAbsoluteError, MeanSquaredError, evaluate
 from flowcean.strategies.incremental import learn_incremental
