+++ examples/coffee_machine/run.py
@@ -1,6 +1,6 @@
 import flowcean.cli
-from flowcean.data.train_test_split import TrainTestSplit
-from flowcean.data.uri import UriDataLoader
+from flowcean.environments.train_test_split import TrainTestSplit
+from flowcean.environments.uri import UriDataLoader
 from flowcean.learners.grpc.learner import GrpcLearner
 from flowcean.metrics import MeanAbsoluteError, MeanSquaredError, evaluate
 from flowcean.strategies.offline import learn_offline
