+++ examples/failure_time_prediction/run.py
@@ -1,8 +1,8 @@
 import logging
 
 import flowcean.cli
-from flowcean.data.train_test_split import TrainTestSplit
-from flowcean.data.uri import UriDataLoader
+from flowcean.environments.train_test_split import TrainTestSplit
+from flowcean.environments.uri import UriDataLoader
 from flowcean.learners.lightning import LightningLearner, MultilayerPerceptron
 from flowcean.metrics import MeanAbsoluteError, MeanSquaredError, evaluate
 from flowcean.strategies.offline import learn_offline
