+++ examples/linear_data/run.py
@@ -3,8 +3,8 @@ import logging
 import polars as pl
 
 import flowcean.cli
-from flowcean.data.dataset import Dataset
-from flowcean.data.train_test_split import TrainTestSplit
+from flowcean.environments.dataset import Dataset
+from flowcean.environments.train_test_split import TrainTestSplit
 from flowcean.learners.linear_regression import LinearRegression
 from flowcean.metrics import MeanAbsoluteError, MeanSquaredError, evaluate
 from flowcean.strategies.incremental import learn_incremental
