+++ tests/data/test_parquet.py
@@ -7,7 +7,7 @@ import pytest
 from polars.testing import assert_frame_equal
 
 from flowcean.core.environment import NotLoadedError
-from flowcean.data.parquet import ParquetDataLoader
+from flowcean.environments.parquet import ParquetDataLoader
 
 
 class TestParquetDataLoader(unittest.TestCase):
