+++ tests/data/test_yaml.py
@@ -7,7 +7,7 @@ import pytest
 from polars.testing import assert_frame_equal
 
 from flowcean.core.environment import NotLoadedError
-from flowcean.data.yaml import YamlDataLoader
+from flowcean.environments.yaml import YamlDataLoader
 
 
 class TestYamlDataLoader(unittest.TestCase):
